DD-54-refactor: 폴더 이동
