+++ src/components/Gnb/GnbButton.tsx
@@ -1,7 +1,7 @@
 import GnbUiButton from '@/components/Gnb/GnbUiButton';
 import NotificationButton from '@/components/Gnb/NotificationButton';
-import LoginSignUp from '@/components/LoginSignUp/LoginSignUp';
 import ModalCustom from '@/components/Modal/ModalCustom';
+import LoginSignUp from '@/components/pageComponents/LoginSignUp/LoginSignUp';
 import useModal from '@/hooks/useModal';
 import { IconCloseBlack } from '@/lib/utils/Icons';
