+++ src/components/pageComponents/ShopDetail/ShopHeader/HeaderImage.tsx
