+++ src/components/pageComponents/ShopDetail/ShopHeader/HeaderInfo.tsx
