+++ src/pages/index.tsx
@@ -3,9 +3,9 @@ import { useRouter } from 'next/router';
 import { useRecoilValue } from 'recoil';
 import Dropdown from '@/components/Dropdown/Dropdown';
 import Filter from '@/components/Filter/Filter';
+import SearchPage from '@/components/pageComponents/SearchPage/searchPage';
 import Pagination from '@/components/Pagination/Pagination';
 import PostCard from '@/components/Post/PostCard';
-import SearchPage from '@/components/SearchPage/searchPage';
 import noticeAPI from '@/lib/api/noticeAPI';
 import useMediaQuery from '@/lib/utils/useMediaQuery';
 import keywordDataState from '@/recoil/atoms/searchAtom';
