+++ src/pages/shops/[shop_id]/index.tsx
@@ -2,8 +2,8 @@ import { GetServerSideProps } from 'next';
 import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { useRecoilValue } from 'recoil';
-import NoticesList from '@/components/ShopDetail/NoticesList';
-import ShopHeaderSection from '@/components/ShopDetail/ShopHeader/ShopHeaderSection';
+import NoticesList from '@/components/pageComponents/ShopDetail/NoticesList';
+import ShopHeaderSection from '@/components/pageComponents/ShopDetail/ShopHeader/ShopHeaderSection';
 import noticeAPI from '@/lib/api/noticeAPI';
 import shopAPI from '@/lib/api/shopAPI';
 import { userState } from '@/recoil/atoms/AuthAtom';
