+++ src/styles/globals.css
@@ -78,11 +78,9 @@ select {
     outline: none;
   }
 }
-
 input::-ms-clear {
   display: none;
 }
-
 input[type='number']::-webkit-inner-spin-button,
 input[type='number']::-webkit-outer-spin-button {
   -webkit-appearance: none;
